fix(enrichment_extraction): Cloning Transformations and Extractions
Fixed an issue where clicking update harvest definitions on a transformation definition
or an extraction definition would cause a running harvest to crash.

This was because we were deleting the clone rather than updating it in place,
causing the reference between a running job and the transformation / extraction definition
to be broken.