feat(optimize_job_data): Atomic Counter Updates
Discovered a race condition using the active record increment method (as it increments the count in memory) so replaced it with
an atomic increment in the database. I had thought initially that increment was doing that but turns out it is not.

Fixed an issue when scheduling enrichments after using an existing extraction wouldn't extract any new data.