conf(optimize_job_data): Brakeman
Ignore possible SQL injection as this is a false positive.
I am using define method to make the class cleaner and it does not take user inputted data.
It is also being done the way that it is so that the update is atomic so that it can handle being run in parallel by multiple sidekiq processes.