refactor(optimize_job_data): Job Table
Improve how we know if we have a report or not to be showed on the table.
If we have a report we want to show it and if we don't have a report but a block has been selected as part of a pipeline to run we want to see that it exists in the table.