fix(enrichment_attribute_issue): Sort the enrichment files so that the pages refer to the correct document
The filesystem order cannot be relied on as different filesystems order files differently. We are essentially treating the file system as an array of pages and needing the index to be the same as the page number, so doing something like @documents[10] needs to give page the 10th page. Without sorting them this is not necessarily the case and you can get back a different page than what you are expecting.